[clean] モジュールの分割を見直しました。名前を変更しました。
