+++ src/main.rs
@@ -1,7 +1,8 @@
-mod common;
+mod discord;
 mod hiyokoslot;
 
-use crate::common::{channel, command};
+use crate::discord::channel;
+use crate::hiyokoslot::{command, bingo, bowling, slot};
 use dotenv::dotenv;
 use serenity::{
     async_trait,
@@ -23,9 +24,9 @@ impl EventHandler for Handler {
         if channel::is_target_channel(channel_name) {
             let command_type = command::get_command_type(&msg);
             match command_type {
-                command::CommandTypeId::HiyokoSlot(n) => hiyokoslot::hiyoko_slot(&ctx, &msg, n).await,
-                command::CommandTypeId::HiyokoBingo => hiyokoslot::hiyoko_bingo(&ctx, &msg).await,
-                command::CommandTypeId::HiyokoBowling => hiyokoslot::hiyoko_bowling(&ctx, &msg).await,
+                command::CommandTypeId::HiyokoSlot(n) => slot::hiyoko_slot(&ctx, &msg, n).await,
+                command::CommandTypeId::HiyokoBingo => bingo::hiyoko_bingo(&ctx, &msg).await,
+                command::CommandTypeId::HiyokoBowling => bowling::hiyoko_bowling(&ctx, &msg).await,
                 command::CommandTypeId::UnknownCommand => println!("This is not target command"),
             };
         }
@@ -50,12 +51,4 @@ async fn main() {
     }
 }
 
-pub mod post {
-    use serenity::{model::channel::Message, prelude::*, utils::MessageBuilder};
-    pub async fn post_message(ctx: &Context, msg: &Message, message_str: String) {
-        let response = MessageBuilder::new().push(message_str).build();
-        if let Err(why) = msg.channel_id.say(&ctx.http, &response).await {
-            println!("Error sending message: {:?}", why);
-        }
-    }
-}
+
